refactor: 가독성을 위해 props type 사용 컴포넌트로 이동
