+++ src/components/common/PeopleCountTag/index.tsx
@@ -1,6 +1,11 @@
 import PeopleCountTagIcon from '@/assets/icons/common/people-count-tag-icon.svg?react';
 import { PeopleCountTagContainer } from '@/components/common/PeopleCountTag/PeopleCountTag.style.ts';
-import { PeopleCountTagProps } from '@/types/props';
+
+export interface PeopleCountTagProps {
+  currentParticipants: number;
+  maxParticipants: number;
+  isClose?: boolean;
+}
 
 const PeopleCountTag = ({
   currentParticipants,
