+++ src/domains/CreatePost/components/CreatePostChildPageLayout/index.tsx
@@ -3,15 +3,17 @@ import ArrowLeftIcon from '@/assets/icons/common/arrow-left-icon.svg?react';
 import { Container, SubTitle } from './Layout.style.ts';
 import { ReactNode } from 'react';
 
+interface Props {
+  subTitle?: string;
+  children: ReactNode;
+  backHandle: () => void;
+}
+
 const CreatePostChildPageLayout = ({
   subTitle,
   children,
   backHandle,
-}: {
-  subTitle?: string;
-  children: ReactNode;
-  backHandle: () => void;
-}) => {
+}: Props) => {
   return (
     <>
       <Header>
