+++ src/domains/CreatePost/components/main/MemberWrap.tsx
@@ -1,7 +1,7 @@
-import { ContentWrapProps } from '@/types/props';
 import NextRadio from '@/domains/CreatePost/components/Radio.tsx';
 import Index from '@/domains/CreatePost/components/ContentWrap';
 import MemberIcon from '@/assets/icons/createPost/member-icon.svg?react';
+import { ContentWrapProps } from '../../types/ContentWrapProps.ts';
 
 const MemberWrap = ({ value, setRegisterDataFunc }: ContentWrapProps) => {
   if (!setRegisterDataFunc) return null;
