+++ src/domains/CreatePost/types/ContentWrapProps.ts
@@ -1,0 +1,7 @@
+import { SetRegisterDataFunc, SetStep } from '@/types';
+
+export interface ContentWrapProps {
+  value: string | { longitude: number; latitude: number };
+  setRegisterDataFunc?: SetRegisterDataFunc;
+  setStep?: SetStep;
+}
