+++ src/domains/Home/components/MoveCurrentLocationButton/index.tsx
@@ -1,11 +1,16 @@
-import { MoveCurrentLocationButtonProps } from '@/types/props/index.ts';
 import { Button } from './MoveCurrentLocation.style.ts';
 
 import ActiveMoveLocationIcon from '@/assets/icons/map/active-move-location-icon.svg?react';
 import NonActiveMoveLocationIcon from '@/assets/icons/map/non-active-move-location-icon.svg?react';
 
-const windowHeight = window.innerHeight;
+interface MoveCurrentLocationButtonProps {
+  moveCurrentLocationFunc: () => void;
+  isActiveMyLocationButton: boolean;
+  isActivePostItem: string | null;
+  postListHeight: number;
+}
 
+const windowHeight = window.innerHeight;
 const MoveCurrentLocationButton = ({
   moveCurrentLocationFunc,
   isActiveMyLocationButton,
